Include RejectMoves unitttest

+++ include/multilevel_bisect/KLFM/KLFM_parallel.hpp
@@ -79,7 +79,7 @@ void find_top_moves(pmondriaan::hypergraph& H,
  * we have to move back vertices from part 0 to part 1 and positive otherwise.
  */
 std::vector<long> reject_unbalanced_moves(long p,
-                                          bulk::queue<long, long, long>& moves_queue,
+                                          bulk::queue<long, long, int>& moves_queue,
                                           std::array<long, 2>& total_weights,
                                           long max_weight_0,
                                           long max_weight_1);
